Use safer types
– std::size_t is platform-dependent, so using it with a portable binary archive does not seem to be such a good idea.

+++ include/founder_graphs/basic_types.hh
@@ -12,7 +12,7 @@
 
 namespace founder_graphs {
 	
-	typedef std::size_t	length_type;
+	typedef std::uint64_t	length_type;
 	constexpr inline auto LENGTH_MAX{std::numeric_limits <length_type>::max()};
 }
